new themes!
added `nord` to the family! we still need to cover gtk themes and icon, but
can consider this a work in progress.

we now use alacritty as main terminal, due to lack of customization on
xfce4-terminal; alacritty is light as its predecessor and can be tweaked on
our needs.

changed fonts as well; iosevka leads space to JetBrainsMono -- a bit wider and
legible.

created a `patcher.sh` script to, given a xresources file, generate files under
theme directory to match configurations on different applications -- like rofi,
alacritty, gtk and so on.