search script
say goodbye to `rofi_zeal` and say hello to, simply, `search`.

this script aims to open a dmenu laucher and present options to
search engines -- via `search_engines.md`. we can search documentations
(manual or tldr pages) on terminal and browse the internet via
duckduckgo, reddit, youtube and so on.